requirements.txt
Flask==0.10.1
gunicorn==19.4.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.1.1
pytesseract==0.1.6
requests==2.9.1
Werkzeug==0.11.4
wheel==0.24.0
